Move instrument helpers to their own module
