+++ satpy/composites/config_loader.py
@@ -28,10 +28,11 @@ import yaml
 from yaml import UnsafeLoader
 
 import satpy
+import satpy._instruments as instru
 from satpy import DataID, DataQuery
 from satpy._config import config_search_paths, get_entry_points_config_dirs, glob_config
 from satpy.dataset.dataid import minimal_default_keys_config
-from satpy.utils import normalize_instrument_name, recursive_dict_update
+from satpy.utils import recursive_dict_update
 
 logger = logging.getLogger(__name__)
 
@@ -268,7 +269,7 @@ def load_compositor_configs_for_sensor(sensor_name: str) -> tuple[dict[str, dict
                 DataID key -> key properties
 
     """
-    config_filename = normalize_instrument_name(sensor_name) + ".yaml"
+    config_filename = instru.normalize_instrument_name(sensor_name) + ".yaml"
     logger.debug("Looking for composites config file %s", config_filename)
     paths = get_entry_points_config_dirs("satpy.composites")
     composite_configs = config_search_paths(
