+++ satpy/modifiers/atmosphere.py
@@ -23,10 +23,10 @@ import dask.array as da
 import numpy as np
 import xarray as xr
 
+import satpy._instruments as instru
 from satpy.modifiers import ModifierBase
 from satpy.modifiers._crefl import ReflectanceCorrector  # noqa
 from satpy.modifiers.angles import compute_relative_azimuth, get_angles, get_satellite_zenith_angle
-from satpy.utils import get_one_instrument_from_attrs, get_pyspectral_instrument_name
 
 logger = logging.getLogger(__name__)
 
@@ -105,8 +105,8 @@ class PSPRayleighReflectance(ModifierBase):
         logger.info("Removing Rayleigh scattering with atmosphere '%s' and "
                     "aerosol type '%s' for '%s'",
                     atmosphere, aerosol_type, vis.attrs["name"])
-        sensor = get_pyspectral_instrument_name(
-            get_one_instrument_from_attrs(vis.attrs)
+        sensor = instru.get_pyspectral_instrument_name(
+            instru.get_one_instrument_from_attrs(vis.attrs)
         )
         corrector = Rayleigh(vis.attrs["platform_name"], sensor,
                              atmosphere=atmosphere,
@@ -162,8 +162,8 @@ class PSPAtmosphericalCorrection(ModifierBase):
         satz = satz.data  # get dask array underneath
 
         logger.info("Correction for limb cooling")
-        sensor = get_pyspectral_instrument_name(
-            get_one_instrument_from_attrs(band.attrs)
+        sensor = instru.get_pyspectral_instrument_name(
+            instru.get_one_instrument_from_attrs(band.attrs)
         )
         corrector = AtmosphericalCorrection(band.attrs["platform_name"],
                                             sensor)
